Change integer to float for exchanges variable

+++ app/controllers/exchange/four_controller.rb
@@ -21,7 +21,7 @@ class Exchange::FourController < ExchangesController
 
       @rate = response_obj['rates']["#{@rate_to}"]
       @rates = response_obj['rates']
-      @exchanges = @input.to_i * @rate.to_i
+      @exchanges = @input.to_f * @rate.to_f
 
       unless @exchanges == 0
         render :index
